fix: ifdefにも正式に対応。さらにヘルパー部分も別ファイルに分離してみた
