+++ ABC460/D.cpp
@@ -1,143 +1,10 @@
 ﻿#ifndef ___INCLUDED_MAIN___
 #define ___INCLUDED_MAIN___
 
-#include <cstdint>
-#include <iostream>
-#include <sstream>
-#include <string>
-
-/**
-    @brief	Atcoderの解答を行うのに便利なクラス
- */
-class AtcoderSolveHelper
-{
-  private:
-    //----------- 編集エリア -----------------------------
-    /**
-        @brief	実際に問題を解く関数
-        @note   マルチケースの場合はケースごとに呼ばれます
-     */
-    void Solve();
-
-    //----------- 以下編集の必要なし ----------------------
-
-  public:
-    /**
-        @brief	各種処理の起点になる窓口
-     */
-    void Run()
-    {
-        int64_t numTestCases = 1;
-#ifdef DISABLE_MULTICASE
-        In() >> numTestCases;
-#endif
-        while (numTestCases--)
-        {
-            Solve();
-        }
-
-        std::cout << Out().str();
-    }
-
-  private:
-    /**
-        @brief	対象とするコンテナの各要素に入力を受け取る
-
-        @tparam	TargetContainerType 対象とするコンテナの型
-        @param[in]	targetContainer 対象とするコンテナ
-     */
-    template <class TargetContainerType>
-    void EachInput(TargetContainerType &targetContainer)
-    {
-        for (auto &&target : targetContainer)
-        {
-            In() >> target;
-        }
-    }
-
-    /**
-        @brief	入力ストリームを取得
-
-        @return 入力ストリーム
-     */
-    std::istream &In()
-    {
-        return std::cin;
-    }
-
-    /**
-        @brief	出力ストリームを取得
-
-        @return 出力ストリーム
-     */
-    std::stringstream &Out()
-    {
-        return m_outStream;
-    }
-
-    /**
-        @brief	エラー出力ストリームを取得
-
-        @return エラー出力ストリーム
-     */
-    std::ostream &Error()
-    {
-        return std::cerr;
-    }
-
-    /**
-        @brief	Yesと1行出力する
-     */
-    void Yes()
-    {
-        Out() << "Yes" << std::endl;
-    }
-
-    /**
-        @brief	Noと1行出力する
-     */
-    void No()
-    {
-        Out() << "No" << std::endl;
-    }
-
-    /**
-        @brief	渡されたbool値に応じてYesかNoを出力する
-
-        @param[in]	isYes Yesを出力するかどうか
-     */
-    void YesNo(bool isYes)
-    {
-        if (isYes)
-        {
-            Yes();
-        }
-        else
-        {
-            No();
-        }
-    }
-
-  private:
-    std::stringstream m_outStream; //!< 出力を一時的に貯めておけるストリーム
-};
-
-/**
-    @brief	エントリポイント
-
-    @return ステータスコード(atcoderにおいては基本0)
- */
-int main()
-{
-    AtcoderSolveHelper solver;
-    solver.Run();
-    return 0;
-}
-
-//----------------------編集スペースここから--------------------------
-#define DISABLE_MULTICASE //!< マルチケース用スイッチ：コメントアウトするとマルチケースになる
+//#define ENABLE_MULTICASE //!< マルチケース用スイッチ：マルチケースの場合はコメント解除
 
 #include "../module/Grid2D.cpp"
+#include "../module/Helper.cpp"
 
 void AtcoderSolveHelper::Solve()
 {
@@ -226,7 +93,5 @@ void AtcoderSolveHelper::Solve()
         Out() << std::endl;
     }
 }
-
-//----------------------編集スペースここまで--------------------------
 
 #endif //___INCLUDED_MAIN___
